requirements.txt
anyio
asgiref
async-timeout
autopep8
certifi
click
Django
django-ckeditor
django-js-asset
django-redis
djangorestframework
exceptiongroup
flake8
googletrans
h11
h2
hpack
httpcore
httpx
hyperframe
idna
iniconfig
mccabe
mypy-extensions
packaging
pathspec
platformdirs
pluggy
pycodestyle
pyflakes
pytest
pytest-django
redis
sniffio
sqlparse
tomli
typing_extensions
psycopg2
psycopg2-binary